requirements.txt
numpy
matplotlib
pandas
torch
torchvision
scipy
tqdm
faiss-gpu
